Moved EmptyGeometry plugin to MarlinCorePlugins module library
